TT-827: Run acceptance tests on pre-commit and default to localhost
The acceptance tests should be run on pre-commit in order to verify that
the application satisfies its contract with the other related
applications.

The locations for Verify-Service-Provider and Compliance-Tool default to
localhost because the tests can be run by anyone anywhere.

On our CI environments and individual development machines we can use
environment variables to run against PAAS and thus avoid spinning up
services locally.

Authors: @andy-paine @tunylund