bau: Use a GITHUB_TOKEN for github API calls (if available)
Some of our Travis builds have been failing because Travis' IP has hit
the github rate limit. By authenticating to the API with a GITHUB_TOKEN
the rate limit is upped to 5000 requests for the user that owns the
token, which should be plenty.

The code's a bit messy because (a) docker (b) I want it to work if
GITHUB_TOKEN is not set.