bau: Fix template lookup on PaaS
In the latest version of the nodejs buildpack on PaaS the location of
the node_modules directory has changed. Previously the code assumed it
would be in `./node_modules` relative to the current working directory.
It's now in `/deps/0/node_modules` on PaaS.

This causes an issue where the views can't find the govuk_template
template.

Instead of relying on the node script being run from a certain directory
we can use [require.resolve](https://nodejs.org/api/modules.html#modules_require_resolve_request_options)
to look up the path to the module using node's resolution algorithm. We
then have to be a bit hacky and go up a couple of directories to get the
`views` dir that we were configuring nunjucks to use before.

Hopefully this should fix the PaaS apps.