bau: Fix asset loading on PaaS
As described in 8612941 the location of node_modules is different in the
latest nodejs buildpack used by PaaS. We'd fixed one of these issues
using require.resolve, but we'd missed the same issue 5 lines further
down.

I've done a `git grep` for `node_modules` and I'm pretty sure this is
the last one.