BAU: Fix the acceptance test
The acceptance tests are written againts the compliance tool. The tests use the
latest released version of VSP which now by default uses non-matching.
However, compliance tool does not support non-matching. This is to add
matching config for VSP when running the acceptance test. In future we should revisit
to see whether we want to still use the compliance tool for this scenario.
This should make the build green.