HUB-813: Give app container a new entrypoint
It looks like the latest version of the base image being used has
changed it's entry point, and also no longer includes bash.

This change to invoke sh as the entry point when putting the app
container to sleep, as well as invoking the tests in travis.

Also add .dockerignore to ensure that stale node_modules aren't included
at build time which is just confusing.